+++ Editor/StateGraph/Nodes/SubStateNodeView.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using Nonatomic.VSM2.NodeGraph;
 using Nonatomic.VSM2.StateGraph;
-using Nonatomic.VSM2.StateGraph.States;
-using UnityEditor;
 using UnityEditor.Experimental.GraphView;
-using UnityEngine;
 using UnityEngine.UIElements;
 
 namespace Nonatomic.VSM2.Editor.StateGraph.Nodes
@@ -28,7 +24,6 @@ namespace Nonatomic.VSM2.Editor.StateGraph.Nodes
 			ColorizeTitle();
 			AddTitleLabel();
 			AddTitleIcon();
-			AddStateMachineOpenButton();
 			AddProgressBar();
 			AddInputPorts(inputContainer);
 			AddOutputPorts(outputContainer);
@@ -42,21 +37,6 @@ namespace Nonatomic.VSM2.Editor.StateGraph.Nodes
 			AddGlowBorder();
 			CheckCustomWidth();
 			UpdatePosition();
-		}
-
-		private void AddStateMachineOpenButton()
-		{
-			var openButton = new Button(() =>
-			{
-				var substate = (BaseSubStateMachineState) NodeModel.State;
-				ModelSelection.ActiveModel = Application.isPlaying 
-					? substate.SubStateMachine.Model 
-					: substate.Model;
-			});
-			openButton.text = "Open";
-			openButton.name = "open-button";
-
-			TitleContainer.Add(openButton);
 		}
 
 		public override void Update()